Check files in directory for retractions

Check all HTML, DOCX, PDF, and RTF files in a directory for updates to referenced DOIs.

retractcheck_dir(path, ...)

Arguments

path
Path to directory to check

...
Arguments passed on to retractcheck

database

Character. Abbreviation of the databases to search (or for openretractions.com). # #' and rw for # #' retractiondatabase.com). Note that in the absence of an API, # #' searching retractiondatabase.com is rather slow.

return

Character. If all, all DOIs are queried and all results are returned; if unique, the DOIs are queried in the order specified until either a correction or retraction notice is found or all databases have been queried.

Value

retractcheck dataframe with filenames

Examples

# \donttest{
  retractcheck_dir(path = '.')

#> 
#> HOORAY *<(:)

#> None of the DOIs mentioned have indexed retractions or corrections.

#>            file                              doi         database update_type
#> 1     check_doi                 10.1002/job.1787 open_retractions  correction
#> 2     check_doi 10.1111/j.1365-2044.2012.07128.x open_retractions  None found
#> 3      find_doi 10.1111/j.1365-2044.2012.07128.x open_retractions  None found
#> 4      find_doi 10.1111/j.1365-2044.2012.07128.x open_retractions  None found
#> 5      query_or                 10.1002/job.1787 open_retractions  correction
#> 6      query_or                 10.1002/job.1787 open_retractions  correction
#> 7      query_or                 10.1002/job.1858 open_retractions  None found
#> 8  retractcheck                 10.1002/job.1787 open_retractions  correction
#> 9  retractcheck 10.1111/j.1365-2044.2012.07128.x open_retractions  None found
#> 10 retractcheck                 10.1002/job.1787 open_retractions  correction
#> 11 retractcheck 10.1111/j.1365-2044.2012.07128.x open_retractions  None found
#> 12 retractcheck                 10.1002/job.1858 open_retractions  None found
#> 13 retractcheck                 10.1002/job.1787 open_retractions  correction
#> 14 retractcheck 10.1111/j.1365-2044.2012.07128.x open_retractions  None found
#> 15 retractcheck                 10.1002/job.1787 open_retractions  correction
#> 16 retractcheck 10.1111/j.1365-2044.2012.07128.x open_retractions  None found
#> 17 retractcheck                 10.1002/job.1858 open_retractions  None found
#>    retracted       update_doi       publisher
#> 1      FALSE 10.1002/job.1858 Wiley-Blackwell
#> 2         NA             <NA>            <NA>
#> 3         NA             <NA>            <NA>
#> 4         NA             <NA>            <NA>
#> 5      FALSE 10.1002/job.1858 Wiley-Blackwell
#> 6      FALSE 10.1002/job.1858 Wiley-Blackwell
#> 7         NA             <NA>            <NA>
#> 8      FALSE 10.1002/job.1858 Wiley-Blackwell
#> 9         NA             <NA>            <NA>
#> 10     FALSE 10.1002/job.1858 Wiley-Blackwell
#> 11        NA             <NA>            <NA>
#> 12        NA             <NA>            <NA>
#> 13     FALSE 10.1002/job.1858 Wiley-Blackwell
#> 14        NA             <NA>            <NA>
#> 15     FALSE 10.1002/job.1858 Wiley-Blackwell
#> 16        NA             <NA>            <NA>
#> 17        NA             <NA>            <NA>
#>                                                                                                                                                                     title
#> 1  Erratum: Cognitive and affective identification: Exploring the links between different forms of social identification and personality with work attitudes and behavior
#> 2                                                                                                                                                                    <NA>
#> 3                                                                                                                                                                    <NA>
#> 4                                                                                                                                                                    <NA>
#> 5  Erratum: Cognitive and affective identification: Exploring the links between different forms of social identification and personality with work attitudes and behavior
#> 6  Erratum: Cognitive and affective identification: Exploring the links between different forms of social identification and personality with work attitudes and behavior
#> 7                                                                                                                                                                    <NA>
#> 8  Erratum: Cognitive and affective identification: Exploring the links between different forms of social identification and personality with work attitudes and behavior
#> 9                                                                                                                                                                    <NA>
#> 10 Erratum: Cognitive and affective identification: Exploring the links between different forms of social identification and personality with work attitudes and behavior
#> 11                                                                                                                                                                   <NA>
#> 12                                                                                                                                                                   <NA>
#> 13 Erratum: Cognitive and affective identification: Exploring the links between different forms of social identification and personality with work attitudes and behavior
#> 14                                                                                                                                                                   <NA>
#> 15 Erratum: Cognitive and affective identification: Exploring the links between different forms of social identification and personality with work attitudes and behavior
#> 16                                                                                                                                                                   <NA>
#> 17                                                                                                                                                                   <NA>
#>    published_original published_update update_delay
#> 1          2013-02-26       2017-05-28    1552 days
#> 2                <NA>             <NA>      NA days
#> 3                <NA>             <NA>      NA days
#> 4                <NA>             <NA>      NA days
#> 5          2013-02-26       2017-05-28    1552 days
#> 6          2013-02-26       2017-05-28    1552 days
#> 7                <NA>             <NA>      NA days
#> 8          2013-02-26       2017-05-28    1552 days
#> 9                <NA>             <NA>      NA days
#> 10         2013-02-26       2017-05-28    1552 days
#> 11               <NA>             <NA>      NA days
#> 12               <NA>             <NA>      NA days
#> 13         2013-02-26       2017-05-28    1552 days
#> 14               <NA>             <NA>      NA days
#> 15         2013-02-26       2017-05-28    1552 days
#> 16               <NA>             <NA>      NA days
#> 17               <NA>             <NA>      NA days

# }