Check html file for retractions

Check a html file for retractions.

retractcheck_html(path, ...)

Arguments

path
Path to html file to check

...
Arguments passed on to retractcheck

database

Character. Abbreviation of the databases to search (or for openretractions.com). # #' and rw for # #' retractiondatabase.com). Note that in the absence of an API, # #' searching retractiondatabase.com is rather slow.

return

Character. If all, all DOIs are queried and all results are returned; if unique, the DOIs are queried in the order specified until either a correction or retraction notice is found or all databases have been queried.

Value

retractcheck dataframe without filenames

Examples

# \donttest{
  retractcheck_html(system.file("extdata", "manuscript.html", package = "retractcheck"))

#> 
#> HOORAY *<(:)

#> None of the DOIs mentioned have indexed retractions or corrections.

#>                                doi         database update_type retracted
#> 1               10.1038/nmeth.3288 open_retractions  None found        NA
#> 2 10.1111/j.1745-6924.2008.00079.x open_retractions  None found        NA
#> 3                 10.1037/a0023347 open_retractions  None found        NA
#> 4              10.7554/eLife.23383 open_retractions  None found        NA
#> 5         10.1177/0956797614567341 open_retractions  None found        NA
#> 6         10.1177/1745691614528520 open_retractions  None found        NA
#> 7     10.1126/scitranslmed.aaf5027 open_retractions  None found        NA
#> 8            10.1126/scitranslmed. open_retractions  None found        NA
#> 9     10.1373/clinchem.2017.271965 open_retractions  None found        NA
#>   update_doi publisher title published_original published_update update_delay
#> 1         NA        NA    NA                 NA               NA           NA
#> 2         NA        NA    NA                 NA               NA           NA
#> 3         NA        NA    NA                 NA               NA           NA
#> 4         NA        NA    NA                 NA               NA           NA
#> 5         NA        NA    NA                 NA               NA           NA
#> 6         NA        NA    NA                 NA               NA           NA
#> 7         NA        NA    NA                 NA               NA           NA
#> 8         NA        NA    NA                 NA               NA           NA
#> 9         NA        NA    NA                 NA               NA           NA

# }